+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `rg-isearch-current-file` is now bound in `isearch-mode-map` using
   user-prefixed `n`.
+- Capturing (project) tasks now uses local variables
+  `wal/org-capture-tasks` and `wal/org-capture-tasks-heading` to find
+  the file and heading using new function
+  `wal/org-captre-find-location`.
 
 ### Changed